requirements.txt
Django==5.0.2
djangorestframework==3.14.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-experimental==0.0.52
langchain-openai==0.0.8
numpy==1.26.4
openai==1.12.0
pandas==2.2.1
 